requirements.txt
alembic==1.12.1
autopep8==2.0.4
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
img2pdf==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pikepdf==8.7.1
Pillow==10.1.0
pycodestyle==2.11.1
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.1
